+++ js-miniapp-bridge/src/common-bridge.ts
@@ -284,7 +284,10 @@ export class MiniAppBridge {
   }
 
   /**
-   * Get access token from native hostapp.
+   * Associating getAccessToken function to MiniAppBridge object.
+   * This function returns access token details from the host app.
+   * (provided the rakuten.miniapp.user.ACCESS_TOKEN is allowed by the user)
+   * It returns error info if user had denied the custom permission
    */
   getAccessToken() {
     return new Promise<AccessTokenData>((resolve, reject) => {