+++ js-miniapp-sample/src/pages/tests/auth-token.test.js
@@ -3,7 +3,10 @@ import React from 'react';
 import { render, cleanup, screen } from '@testing-library/react';
 
 import { wrapTheme } from '../../tests/test-utils';
-import AuthToken from '../auth-token';
+import { AuthToken } from '../auth-token';
+import '@testing-library/jest-dom/extend-expect';
+
+jest.mock('axios');
 
 beforeEach(() => {
   render(wrapTheme(<AuthToken />));