+++ js-miniapp-sample/src/services/user/actions.js
@@ -1,4 +1,4 @@
-import MiniApp from 'js-miniapp-sdk';
+import MiniApp, { AccessTokenData } from 'js-miniapp-sdk';
 
 import {
   REQUEST_USER_NAME_SUCCESS,
@@ -7,6 +7,8 @@ import {
   REQUEST_PROFILE_PHOTO_FAILURE,
   REQUEST_CONTACT_LIST_SUCCESS,
   REQUEST_CONTACT_LIST_FAILURE,
+  REQUEST_ACCESS_TOKEN_SUCCESS,
+  REQUEST_ACCESS_TOKEN_FAILURE,
 } from './types';
 
 import { Contact } from 'js-miniapp-sdk';
@@ -14,6 +16,7 @@ import { Contact } from 'js-miniapp-sdk';
 type UserNameSuccessAction = { type: String, userName: string };
 type ProfilePhotoSuccessAction = { type: String, url: string };
 type ContactListSuccessAction = { type: String, contacts: Contact[] };
+type AccessTokenSuccessAction = { type: String, token: AccessTokenData };
 
 const requestUserName = (): Function => {
   return (dispatch) => {
@@ -71,9 +74,35 @@ const requestContactList = (): Function => {
   };
 };
 
-export { requestUserName, requestProfilePhoto, requestContactList };
+const requestAccessToken = (): Function => {
+  return (dispatch) => {
+    return MiniApp.user
+      .getAccessToken()
+      .then((token) => {
+        dispatch({
+          type: REQUEST_ACCESS_TOKEN_SUCCESS,
+          token,
+        });
+
+        return Promise.resolve(token);
+      })
+      .catch((_) => {
+        dispatch({
+          type: REQUEST_ACCESS_TOKEN_FAILURE,
+        });
+      });
+  };
+};
+
+export {
+  requestUserName,
+  requestProfilePhoto,
+  requestContactList,
+  requestAccessToken,
+};
 export type {
   UserNameSuccessAction,
   ProfilePhotoSuccessAction,
   ContactListSuccessAction,
+  AccessTokenSuccessAction,
 };